requirements.txt
streamlit==1.36.0 # Frontend framework for building the UI 
pydantic==2.9.0 # Core data validation and settings management. Used for structured data models
openai==1.36.0 # Official OpenAI SDK
pydantic-ai==0.15.0 # The main AI agent framework
python-dotenv==1.0.1 # Loads environment variables from .env
nest_asyncio==1.5.8 # Allows nested event loops, particularly in environments like Jupyter or Streamlit where asyncio is already running
psycopg2-binary==2.9.9 # PostgreSQL adapter for Python. Provides synchronous access to PostgreSQL when needed
asyncpg==0.29.0 # Asynchronous PostgreSQL driver.
requests
pillow
requests
beautifulsoup4
py-lighthouse-audit